TBot/requirements.txt
python-telegram-bot==12.7
pandas==1.5.3
